requirements.txt
python-aiml==0.9.3
Flask==1.1.2
waitress==1.4.3